[no-issue] Added support to scroll-aware list dropdown positioning
If the component was placed on the bottom of a page, the dropdown increased the body size, leading to styling issues (look and feel of the page might be altered due to larger body size).

Regarding the scroll position scenarios the component mimics the behaviour of the built-in <datalist> component:
-if there is not enough room for the list to display below the component, it is displayed above the component.
-if there is enough room for the list to display below the component, it is displayed below the component.

For correct size calculation and to avoid dependency on properly set max-size CSS property the .autocomplete-list.hidden css class was updated to use visibility: hidden instead of display:none. The former results in correct bounding client rect when the list is hidden.